modules/rag-service/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.2
    # via anyio
fastapi[standard]==0.115.0
    # via -r requirements.in
fastapi-cli[standard]==0.0.5
    # via fastapi
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via phidata
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via
    #   fastapi
    #   ollama
    #   phidata
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.1
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==2.1.2
    # via pgvector
ollama==0.3.3
    # via -r requirements.in
pgvector==0.3.5
    # via -r requirements.in
phidata==2.4.42
    # via -r requirements.in
psycopg[binary]==3.2.3
    # via -r requirements.in
psycopg-binary==3.2.3
    # via psycopg
pydantic==2.9.2
    # via
    #   fastapi
    #   phidata
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.5.2
    # via phidata
pygments==2.18.0
    # via rich
pypdf==5.0.1
    # via -r requirements.in
python-consul==1.1.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   phidata
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.12
    # via fastapi
pyyaml==6.0.2
    # via
    #   phidata
    #   uvicorn
requests==2.32.3
    # via
    #   -r requirements.in
    #   python-consul
rich==13.9.2
    # via
    #   phidata
    #   typer
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-consul
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.35
    # via -r requirements.in
starlette==0.38.6
    # via fastapi
tomli==2.0.2
    # via phidata
typer==0.12.5
    # via
    #   fastapi-cli
    #   phidata
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   phidata
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   pypdf
    #   rich
    #   sqlalchemy
    #   typer
    #   uvicorn
tzdata==2024.2
    # via psycopg
urllib3==2.2.3
    # via requests
uvicorn[standard]==0.31.1
    # via
    #   fastapi
    #   fastapi-cli
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
